+++ master/internal/api_tensorboard.go
@@ -14,10 +14,8 @@ import (
 
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/encoding/protojson"
 
 	petname "github.com/dustinkirkland/golang-petname"
-	"github.com/ghodss/yaml"
 
 	"github.com/pkg/errors"
 
@@ -267,7 +265,7 @@ func (a *apiServer) LaunchTensorboard(
 
 	spec.Base.ExtraEnvVars = uniqEnvVars
 
-	if !usingCustomImage(req) {
+	if !model.UsingCustomImage(req) {
 		spec.Config.Environment.Image = model.RuntimeItem{
 			CPU:  expConf.Environment().Image().CPU(),
 			CUDA: expConf.Environment().Image().CUDA(),
@@ -319,31 +317,6 @@ func (a *apiServer) LaunchTensorboard(
 		Tensorboard: tb,
 		Config:      protoutils.ToStruct(spec.Config),
 	}, err
-}
-
-func usingCustomImage(req *apiv1.LaunchTensorboardRequest) bool {
-	if req.Config == nil {
-		return false
-	}
-
-	configBytes, err := protojson.Marshal(req.Config)
-	if err != nil {
-		return false
-	}
-
-	type DummyImg struct {
-		Image string `json:"image"`
-	}
-	type DummyEnv struct {
-		Environment DummyImg `json:"environment"`
-	}
-	dummy := DummyEnv{
-		Environment: DummyImg{},
-	}
-
-	err = yaml.Unmarshal(configBytes, &dummy)
-
-	return err == nil && dummy.Environment.Image != ""
 }
 
 type tensorboardConfig struct {
